Made requested changes and added patientUtils file
